audit pk_stmts:
Having to cast everything to text slows things down a lot.  But it has to be
done, because:

"PostgreSQL's implicit casting rules depend on context. When comparing text with
integer explicitly (e.g., as part of a query or function), PostgreSQL does not
assume how the types should be cast, so it fails with an error."

Would be very interesting to try doing column type inspection to possibly speed this up.